refactor: parsing관련 함수를 parsingService로 분리
